add some code

+++ paddle/fluid/prim/api/composite_backward/composite_backward_api.h
@@ -1821,7 +1821,7 @@ void pad_grad(const Tensor& input,
     std::vector<int64_t> axes(rank, 0);
     std::vector<int64_t> infer_flags(rank, 1);
     std::vector<int64_t> decrease_axis({});
-    for (int64_t i = 0; i < rank; ++i) {
+    for (size_t i = 0; i < rank; ++i) {
       starts.push_back(static_cast<int>(paddings[2 * i]));
       ends.push_back(static_cast<int64_t>(out_dims[i] - paddings[2 * i + 1]));
       axes.push_back(i);
